+++ vite.config.ts
@@ -7,6 +7,9 @@ import react from '@vitejs/plugin-react';
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  define: {
+    BASE_URL: JSON.stringify(''),
+  },
   test: {
     globals: true,
     environment: 'jsdom',
